+++ shared/src/business/entities/cases/IncompleteEditElectronicPetition.test.ts
@@ -4,7 +4,7 @@ import {
   MAX_FILE_SIZE_MB,
   PARTY_TYPES,
 } from '../EntityConstants';
-import { CaseExternal } from './CaseExternal';
+import { CaseExternal } from './IncompleteEditElectronicPetition';
 import { applicationContext } from '../../test/createTestApplicationContext';
 
 describe('CaseExternal entity', () => {
